putnbr and norminette

+++ ft_memccpy.c
@@ -6,7 +6,7 @@
 /*   By: rgrootho <marvin@codam.nl>                   +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/11/01 16:45:40 by rgrootho      #+#    #+#                 */
-/*   Updated: 2021/03/26 13:30:09 by rgrootho      ########   odam.nl         */
+/*   Updated: 2021/04/26 16:41:18 by rgrootho      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 void	*ft_memccpy(void *dst, const void *src, int c, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
-	while (i < (int)n)
+	while (i < n)
 	{
 		*(unsigned char *)(dst + i) = *(unsigned char *)(src + i);
 		if ((unsigned char)c == *(unsigned char *)(src + i))
