+++ ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: rgrootho <marvin@codam.nl>                   +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/11/05 15:52:47 by rgrootho      #+#    #+#                 */
-/*   Updated: 2021/03/26 13:32:26 by rgrootho      ########   odam.nl         */
+/*   Updated: 2021/04/26 16:46:55 by rgrootho      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,14 +14,14 @@
 
 int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
-	while (i < (int)n)
+	while (i < n)
 	{
 		if (*(unsigned char *)(s1 + i) != *(unsigned char *)(s2 + i))
-			return (*(unsigned char*)(s1 + i)
-				- *(unsigned char*)(s2 + i));
+			return (*(unsigned char *)(s1 + i)
+				- *(unsigned char *)(s2 + i));
 		else
 			i++;
 	}
