+++ ft_strchr.c
@@ -6,7 +6,7 @@
 /*   By: rgrootho <marvin@codam.nl>                   +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/11/07 17:50:09 by rgrootho      #+#    #+#                 */
-/*   Updated: 2021/03/26 13:58:17 by rgrootho      ########   odam.nl         */
+/*   Updated: 2021/04/26 17:06:41 by rgrootho      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,16 +14,16 @@
 
 char	*ft_strchr(const char *s, int c)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
 	{
 		if (s[i] == (char)c)
-			return ((char*)s + i);
+			return ((char *)s + i);
 		i++;
 	}
 	if ((char)c == '\0')
-		return ((char*)s + i);
+		return ((char *)s + i);
 	return (0);
 }
